assembly: support freezing assembly
New property 'Freeze' added to assembly container to freeze any update
of the assembly.

AsmPartGroup and AsmElement has been change to bind to
Part::FeaturePython instead of App::FeaturePython, in order to be able
to hold its own shape when frozen. The new ChildViewProvider feature in
ViewProviderLink is used to display the native shape.