assembly: fix frozen assembly topo naming
Assembly contain does not save shape to file, but retores the shape
from part group when restoring document. The problem occurs when
assigning the Assembly object ID to the shape tag before assigning,
causing the element tracing to skip the part group, which in turn cause
loss of color for frozen assembly.

Fixes realthunder/FreeCAD_Assembly3#1185